extract Camera component
